Refactoring the Tado Classes by creating a common base class.
